pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["hackspace_storage"]

[project]
name = "hackspace-storage"
version = "0.1.0"
dependencies = [
    "Flask ~= 3.1.0",
    "Flask-SQLAlchemy ~= 3.1.1",
    "Flask-Migrate ~= 4.1.0",
    "Flask-WTF ~= 1.2.2",
    "psycopg2-binary ~= 2.9.10",
    "cryptography ~= 45.0.6",
    "pyjwt ~= 2.10.1"
]